Everything working except for pricing model

+++ Assignment2/Order.cs
@@ -5,14 +5,14 @@ namespace Assignment2
     public class Order
     {
         public string SenderId { get; set; }        
-        public long CardNo { get; set; }
+        public int CardNo { get; set; }
         public int Amount { get; set; }
         public float UnitPrice { get; set; }
         public DateTime TimeCreated { get; set; }
         public DateTime TimeFulfilled { get; set; }
         public string ReceiverId { get; set; }
 
-        public Order(string senderId, long cardNo, int amt, float unitPrice)
+        public Order(string senderId, int cardNo, int amt, float unitPrice)
         {
             SenderId = senderId;
             CardNo = cardNo;
